merge wiisxrxDma 并且修正了Cdrom读取的问题

+++ wiisxr/dfsound/externals.h
@@ -81,7 +81,8 @@ typedef struct
  unsigned char  SustainLevel;
  unsigned char  SustainRate;
  unsigned char  ReleaseRate;
- int            EnvelopeVol;
+ float          EnvelopeVol;
+ //int            EnvelopeVol;
 } ADSRInfoEx;
 
 ///////////////////////////////////////////////////////////
